+++ fltk-sys/src/lib.rs
@@ -21,6 +21,7 @@ pub mod tree;
 pub mod valuator;
 pub mod widget;
 pub mod window;
+pub mod surface;
 
 #[cfg(test)]
 mod tests {